requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
actuator-msgs==0.0.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
bond==4.1.2
Brotli==1.1.0
builtin-interfaces==1.2.2
cartographer-ros-msgs==2.0.9002
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
colorlog==6.9.0
composition-interfaces==1.2.2
contourpy==1.3.2
control-msgs==4.8.0
controller-manager==2.51.0
controller-manager-msgs==2.51.0
ctranslate2==4.6.0
cv-bridge==3.2.1
cycler==0.12.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
diagnostic-updater==4.0.6
docopt==0.6.2
domain-coordinator==0.10.0
dwb-msgs==1.1.18
dynamixel-sdk==3.8.4
einops==0.8.1
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
exceptiongroup==1.3.0
fastapi==0.116.1
faster-whisper==1.2.0
ffmpy==0.6.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
gazebo-msgs==3.9.0
generate-parameter-library-py==0.5.0
geometry-msgs==4.9.0
gps-msgs==2.1.1
gradio==5.42.0
gradio_client==1.11.1
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
image-geometry==3.2.1
interactive-markers==2.3.2
irobot-create-common-bringup==2.1.0
irobot-create-msgs==2.1.0
Jinja2==3.1.6
joblib==1.5.1
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
julius==0.2.7
kiwisolver==1.4.9
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lifecycle-msgs==1.2.2
lightning==2.5.3
lightning-utilities==0.15.2
logging-demo==0.20.5
Mako==1.3.10
map-msgs==2.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
message-filters==4.3.8
moveit-msgs==2.2.1
mpmath==1.3.0
multidict==6.6.4
nav-2d-msgs==1.1.18
nav-msgs==4.9.0
nav2-common==1.1.18
nav2-msgs==1.1.18
nav2-simple-commander==1.0.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
object-recognition-msgs==2.0.0
octomap-msgs==2.0.1
omegaconf==2.3.0
onnxruntime==1.22.1
optuna==4.4.0
orjson==3.11.2
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pillow==11.3.0
primePy==1.3
propcache==0.3.2
protobuf==6.32.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-qt-binding==1.1.2
pytorch-lightning==2.5.3
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
requests==2.32.4
resource-retriever==3.1.3
rich==14.1.0
rmw-dds-common==1.6.0
ros-gz-interfaces==0.244.20
ros-ign-interfaces==0.244.20
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2controlcli==2.51.0
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-monitor==1.0.6
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.9
safehttpx==0.1.6
scikit-learn==1.7.1
scipy==1.15.3
scripts==3.9.0
semantic-version==2.10.0
semver==3.0.4
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
sentencepiece==0.2.1
shape-msgs==4.9.0
shellingham==1.5.4
six==1.17.0
slam-toolbox==2.6.10
smclib==4.1.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.43
srdfdom==2.0.7
sros2==0.10.7
starlette==0.47.2
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.14.0
tabulate==0.9.0
teleop-twist-keyboard==2.4.0
tensorboardX==2.6.4
tf-transformations==1.1.0
tf2-geometry-msgs==0.25.15
tf2-kdl==0.25.15
tf2-msgs==0.25.15
tf2-py==0.25.15
tf2-ros-py==0.25.15
tf2-sensor-msgs==0.25.15
tf2-tools==0.25.15
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
tomlkit==0.13.3
topic-monitor==0.20.5
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.1
tqdm==4.67.1
trajectory-msgs==4.9.0
triton==3.4.0
turtlebot3-applications-msgs==1.0.1
turtlebot3-autorace-camera==1.2.2
turtlebot3-autorace-detect==1.2.2
turtlebot3-autorace-mission==1.2.2
turtlebot3-example==2.3.3
turtlebot3-msgs==2.4.0
turtlebot3-teleop==2.3.3
turtlebot4-msgs==1.0.5
turtlebot4-navigation==1.0.5
turtlesim==1.4.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unique-identifier-msgs==2.2.1
urdfdom-py==1.2.1
urllib3==2.5.0
uvicorn==0.35.0
vision-msgs==4.1.1
visualization-msgs==4.9.0
websockets==15.0.1
xacro==2.0.13
yarl==1.20.1
